Update index.html
For the transaction tab again, is it possible that the cards under the transaction history to be a 2x2 grid type? or I wanted to see them side by side. and also can you add additional style for the cards to make it look professional and aesthetic at the same time