Update index.html
Thank you for this revision! I would like to add another revisions again. Please refer on the following:

(1) For the dashboard is it possible that the transaction cards below the dashboard summary to be group based on their type? example forthe credit cards, is it possible that they will grouped in one card section? refer on the image provided

(2) for the Add transaction card, it seems that its card size is not aligned or the same with the others in the tab. Its too small. Please see attached image for what I am talking about

(3) is it possible that the all the "ADD" buttons of each tabs will be located in the upper right side? please refer on the image provided. can you also follow the color of the button on the image provided?