Update index.html
Thank you for this, but I think you misunderstood me again. please refer on the changes below:

(1) Th hovering effects of the cards are now missing for all the tabs

(2) Can you remove the small add transaction button located the lower left side of the add transaction button? refer on the image provided

(3) for the add button, can you make all the add buttons follow the position of the Add Transaction button? which located at the upper right side of the section heading. refer on the image provided

(4) Can you return the format of the cards in the tab of cards, goals and reminders? return it to the previous format before this current code